# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Rin OJ Gateway API
  version: 0.1.0
  description: External HTTP contract for Rin OJ gateway.
servers:
  - url: https://api.rin-oj.local
security:
  - bearerAuth: []
paths:
...
  /v1/admin/users/{userId}/role:
    patch:
      operationId: updateUserRole
      tags: [Admin]
      summary: Change the platform role of a user. Caller must have the admin role.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: User ID or username of the target user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleRequest"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateUserRoleResponse"
        "400":
          description: Invalid role value
        "401":
          description: Missing actor identity
        "403":
          description: Caller is not an admin
        "404":
          description: Target user not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    UpdateUserRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [admin, teacher, student]
          description: The new role to assign to the target user.
    UpdateUserRoleResponse:
      type: object
      required: [userId, role]
      properties:
        userId:
          type: string
        role:
          type: string
          enum: [admin, teacher, student]
...